+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add `Teamtailor::Stage` and `Client#stages` for fetching it (`2179b1`)
+
 ## v0.2.3 - 2020-04-06
 
 - Add `Teamtailor::Company` and `Client#company` for fetching it (`05dde0`)